Create firmware.ino
Redesigned fundamental architecture to avoid interrupt calls. 
The controller now has a modifiable schedule which stores the tasks it needs to do.
The controller runs the next task in the schedule every 20 microseconds.
To ensure proper operation, all tasks must be kept to less than 20 microseconds (this has been tested and is feasible).

Moved each object to separate header files.
Created a header called commands.h which lists a series of macros storing the byte value associated to each scheduled task the controller can perform. 

Rewrote the motor driver so that it now interprets velocities in physical units. 
Will probably change how the motor driver handles the velocities of the actuators into a 
method which is more amenable to feedback.

Created a new Port object which handles the I/O operations. Specifically, it handles the transmission and reception
of 64 byte packets between the device (the teensy) and the host (a C++ program on the NUC). 
Have tested to ensure that the decoding of a packet is achievable within 20 microseconds. (it takes about 14 microseconds to decode
a packet and schedule tasks)

Made some minor formatting and style changes so that the firmware more closely follows the conventions of
Google's c++ style guide. (which was the chosen style)

Plan to move entirely away from the use of floats on the device and instead to pure signed integers .
Any floating point operations could then be handled on the NUC (not sure about this though).