Added Navigator
Implemented a navigator object which takes a given displacement vector and drives the robot to it. 
There is an error of about 5% in the final position for pure x and y movements. This is likely due to inaccuracy in the distance/microstep measure (which needs to be determined in a more accurate manner)
Implemented step counting and the associated message handling.
Rationalized the organisation of the primary driver file.