Limit switch handling
Implemented the firmware side of limit switch handling for the robot. 
Need to finalise limit switch mounting as current temporary mount is unstable. Also, need to measure
the number of microsteps corresponding to the full throw of the actuators and implement a software limit on 
other end based on that maximum throw.